chore(alloy): cleanup alloy evm
