+++ crates/alloy/evm/src/ctx.rs
@@ -1,0 +1,12 @@
+use alloy_primitives::{B256, Bytes};
+
+/// Context for OP block execution.
+#[derive(Debug, Default, Clone)]
+pub struct OpBlockExecutionCtx {
+    /// Parent block hash.
+    pub parent_hash: B256,
+    /// Parent beacon block root.
+    pub parent_beacon_block_root: Option<B256>,
+    /// The block's extra data.
+    pub extra_data: Bytes,
+}
