+++ crates/alloy/evm/src/error.rs
@@ -1,0 +1,24 @@
+use alloc::boxed::Box;
+
+/// Custom errors that can occur during OP block execution.
+#[derive(Debug, thiserror::Error)]
+pub enum OpBlockExecutionError {
+    /// Failed to load cache account.
+    #[error("failed to load cache account")]
+    LoadCacheAccount,
+
+    /// Failed to get Jovian da footprint gas scalar from database.
+    #[error("failed to get da footprint gas scalar from database: {_0}")]
+    GetJovianDaFootprintScalar(Box<dyn core::error::Error + Send + Sync + 'static>),
+
+    /// Transaction DA footprint exceeds available block DA footprint.
+    #[error(
+        "transaction DA footprint exceeds available block DA footprint. transaction_da_footprint: {transaction_da_footprint}, available_block_da_footprint: {available_block_da_footprint}"
+    )]
+    TransactionDaFootprintAboveGasLimit {
+        /// The DA footprint of the transaction to execute.
+        transaction_da_footprint: u64,
+        /// The available block DA footprint.
+        available_block_da_footprint: u64,
+    },
+}
