+++ crates/alloy/evm/src/tx_env.rs
@@ -1,0 +1,15 @@
+use alloy_primitives::Bytes;
+use op_revm::OpTransaction;
+
+/// Trait for OP transaction environments. Allows to recover the transaction encoded bytes if
+/// they're available.
+pub trait OpTxEnv {
+    /// Returns the encoded bytes of the transaction.
+    fn encoded_bytes(&self) -> Option<&Bytes>;
+}
+
+impl<T: revm::context::Transaction> OpTxEnv for OpTransaction<T> {
+    fn encoded_bytes(&self) -> Option<&Bytes> {
+        self.enveloped_tx.as_ref()
+    }
+}
